requirements.txt
# Core dependencies
tweepy>=4.14.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Database
sqlalchemy>=2.0.0

# Configuration
pyyaml>=6.0
python-crontab>=2.7.0

# Text processing
nltk>=3.8
textblob>=0.17.1
spacy>=3.6.0

# Scheduling and async
schedule>=1.2.0
aiohttp>=3.8.0

# Logging and monitoring
loguru>=0.7.0
prometheus-client>=0.17.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
